Created a new class file for Product Listing Adapter.
